refactor: Remove dot pattern
