+++ packages/ui/src/organisms/OutOfStock/OutOfStock.tsx
@@ -3,7 +3,7 @@ import type { ReactNode, FormHTMLAttributes } from 'react'
 
 import Form from '../../molecules/Form'
 
-type OutOfStockBaseProps = {
+export type OutOfStockBaseProps = {
   /**
    * ID to find this component in testing tools (e.g.: cypress,
    * testing-library, and jest).
@@ -22,20 +22,6 @@ export type OutOfStockProps = OutOfStockBaseProps & {
   onSubmit?: (event: React.FormEvent<HTMLFormElement>) => void
 } & FormHTMLAttributes<HTMLFormElement>
 
-export type OutOfStockTitleProps = {
-  /**
-   * Attribute used for polymorphic component.
-   */
-  as?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p'
-} & OutOfStockBaseProps
-
-export type OutOfStockMessageProps = {
-  /**
-   * Attribute used for polymorphic component.
-   */
-  as?: 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p' | 'div' | 'span'
-} & OutOfStockBaseProps
-
 const OutOfStock = ({
   testId = 'store-out-of-stock',
   children,
@@ -49,32 +35,5 @@ const OutOfStock = ({
     </section>
   )
 }
-
-const Title = ({
-  as: TitleComponent = 'h2',
-  testId = 'store-out-of-stock-title',
-  children,
-}: OutOfStockTitleProps) => {
-  return (
-    <TitleComponent data-out-of-stock-title data-testid={testId}>
-      {children}
-    </TitleComponent>
-  )
-}
-
-const Message = ({
-  as: MessageComponent = 'p',
-  testId = 'store-out-of-stock-message',
-  children,
-}: OutOfStockMessageProps) => {
-  return (
-    <MessageComponent data-out-of-stock-message data-testid={testId}>
-      {children}
-    </MessageComponent>
-  )
-}
-
-OutOfStock.Title = Title
-OutOfStock.Message = Message
 
 export default OutOfStock
