+++ packages/ui/src/organisms/OutOfStock/OutOfStockMessage.tsx
@@ -1,0 +1,22 @@
+import React from 'react'
+
+import type { OutOfStockBaseProps } from './OutOfStock'
+
+export type OutOfStockMessageProps = {
+  /**
+   * Attribute used for polymorphic component.
+   */
+  as?: 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p' | 'div' | 'span'
+} & OutOfStockBaseProps
+
+export const OutOfStockMessage = ({
+  as: MessageComponent = 'p',
+  testId = 'store-out-of-stock-message',
+  children,
+}: OutOfStockMessageProps) => {
+  return (
+    <MessageComponent data-out-of-stock-message data-testid={testId}>
+      {children}
+    </MessageComponent>
+  )
+}
