+++ packages/ui/src/organisms/OutOfStock/index.tsx
@@ -1,6 +1,6 @@
 export { default } from './OutOfStock'
-export type {
-  OutOfStockProps,
-  OutOfStockMessageProps,
-  OutOfStockTitleProps,
-} from './OutOfStock'
+export { OutOfStockMessage } from './OutOfStockMessage'
+export { OutOfStockTitle } from './OutOfStockTitle'
+export type { OutOfStockProps } from './OutOfStock'
+export type { OutOfStockMessageProps } from './OutOfStockMessage'
+export type { OutOfStockTitleProps } from './OutOfStockTitle'
