+++ packages/ui/src/organisms/OutOfStock/stories/OutOfStock.mdx
@@ -11,8 +11,8 @@ import OutOfStock from '../OutOfStock'
 The `OutOfStock` uses the [Compound Component](https://kentcdodds.com/blog/compound-components-with-react-hooks) pattern, its components are:
 
 - `OutOfStock`: the form wrapper of the out of stock component with title and message;
-- `OutOfStock.Title`: the main title component for out of stock component;
-- `OutOfStock.Message`: the message component for out of stock component;
+- `OutOfStockTitle`: the main title component for out of stock component;
+- `OutOfStockMessage`: the message component for out of stock component;
 
 ## Props
